refactor(hdfs): restructure HdfsBackend and introduce HdfsCore
